feat : Add user check when delete booking table's data in api_booking.py.

+++ controller/api_booking.py
@@ -12,15 +12,12 @@ cnxpool = controller.db_conncetion.db_connection_pool()
 @api_booking.route("/api/booking", methods = ["GET", "POST", "DELETE"])
 def booking():
     try:
-        print("backend")
         cnx = cnxpool.get_connection()
         mycursor = cnx.cursor(dictionary = True)
         cookie = request.cookies.get("Set-Cookie")
         memberInfo = decode_token(cookie)
         if(request.method == "GET"):
-            print("GET")
             if(memberInfo == None):
-                print("GET, 403")
                 return jsonify(
                     {
                         "error" : True,
@@ -28,16 +25,10 @@ def booking():
                     }
                 ), 403
             else:
-                print(memberInfo)
-                print("GET, 200")
                 sql = "select booking.id, booking.attraction_id, booking.date, booking.time, booking.price, attraction.name, attraction.address, attraction.images from booking inner join attraction on booking.attraction_id = attraction.id where booking.user_id = %s"
-                print("GET, 200-1")
                 value = (memberInfo["id"], )
-                print("GET, 200-2")
                 mycursor.execute(sql, value)
-                print("GET, 200-3")
                 bookingData = mycursor.fetchall()
-                print(bookingData)
                 bookingInfoResponseToFrontEnd = []
                 for bookingInfo in bookingData:
                     attraction = {
@@ -60,24 +51,15 @@ def booking():
                     }
                     ), 200
         elif(request.method == "POST"):
-            print("POST")
             # cookie = request.cookies.get("Set-Cookie")
-            print(cookie)
             # get user_id from memberInfo after decodeJWT
             memberInfo = decode_token(cookie)
-            print(memberInfo)
             bookingData = request.get_json()
-            print(bookingData)
             checkAttractionIdFormat = regexDigitalNumber(bookingData["attractionId"])
-            print(checkAttractionIdFormat)
             checkDateFormat = regexDate(bookingData["date"])
-            print(checkDateFormat)
             checkTimeFormat = regexTime(bookingData["time"])
-            print(checkTimeFormat)
             checkPriceFormat = regexDigitalNumber(bookingData["price"])
-            print(checkPriceFormat)
             if(memberInfo == None):
-                print("403")
                 return jsonify(
                     {
                         "error" : True,
@@ -90,7 +72,6 @@ def booking():
                 or checkTimeFormat == False 
                 or checkPriceFormat == False
             ):
-                print("400")
                 return jsonify(
                     {
                         "error" : True,
@@ -104,7 +85,6 @@ def booking():
                 and checkTimeFormat == True 
                 and checkPriceFormat == True
                 ):
-                    print("200")
                     sql = "insert into booking (attraction_id, date, time, price, user_id) values(%s, %s, %s, %s, %s)"
                     value = (bookingData["attractionId"], bookingData["date"], bookingData["time"], bookingData["price"], memberInfo["id"])
                     mycursor.execute(sql, value)
@@ -116,7 +96,6 @@ def booking():
                         }
                     ), 200
             else:
-                print("500")
                 return jsonify(
                     {
                         "error" : True,
@@ -132,11 +111,10 @@ def booking():
                     }
                 ), 403
             elif(memberInfo):
-                print("Delete, 200")
+                userId = memberInfo["id"]
                 bookingWantToDelete = request.get_json()
-                print(bookingWantToDelete["bookingWantToDelete"])
-                sql = "delete from booking where id = %s"
-                value = (bookingWantToDelete["bookingWantToDelete"], )
+                sql = "delete from booking where id = %s and user_id = %s"
+                value = (bookingWantToDelete["bookingWantToDelete"], userId)
                 mycursor.execute(sql, value)
                 cnx.commit()
                 return jsonify(
@@ -145,7 +123,6 @@ def booking():
                     }
                 ), 200
     except:
-        print("except")
         return jsonify(
             {
                 "error" : True,
@@ -153,6 +130,5 @@ def booking():
             }
         ), 500
     finally:
-        print("finally")
         mycursor.close()
         cnx.close()
